Requested logic flow changes

+++ src/datoso/helpers/__init__.py
@@ -103,10 +103,6 @@ class FileUtils:
                 shutil.copy(origin, destination)
         except shutil.SameFileError:
             pass
-        except IsADirectoryError:
-            with suppress(FileNotFoundError):
-                shutil.rmtree(destination)
-            shutil.copytree(origin, destination)
         except FileNotFoundError:
             raise FileNotFoundError(f"File {origin} not found.")
 
@@ -119,15 +115,12 @@ class FileUtils:
     @staticmethod
     def remove(path):
         """ Remove file or folder. """
-        try:
-            if os.path.isdir(path):
-                FileUtils.remove_folder(path)
-            else:
-                os.unlink(path)
-        except IsADirectoryError:
+        if not os.path.exists(path):  
+            return
+        if os.path.isdir(path):
             FileUtils.remove_folder(path)
-        except FileNotFoundError:
-            pass
+        else:
+            os.unlink(path)
 
     @staticmethod
     def parse_folder(path) -> str:
